requirements.txt
Django==4.1.2
psycopg2==2.9.4
gunicorn==20.1.0
djangorestframework==3.14.0
requests